refactor data mapping, esp totals
