+++ openfecwebapp/views.py
@@ -1,7 +1,11 @@
 from flask import render_template
-from openfecwebapp.data_mappings import (type_map, 
-map_candidate_table_values, map_committee_table_values, 
-map_candidate_page_values, add_committee_data, generate_pagination_values)
+from openfecwebapp.data_prep.candidates import (map_candidate_table_values,
+map_candidate_page_values)
+from openfecwebapp.data_prep.committees import (map_committee_table_values,
+map_committee_page_values)
+from openfecwebapp.data_prep.shared import generate_pagination_values
+from openfecwebapp.data_prep.financial_summaries import add_committee_financial_data
+
 from werkzeug.exceptions import abort
 
 def render_search_results(results, query):
@@ -40,6 +44,13 @@ def render_table(data_type, results, params, url):
 
     return render_template(data_type + '.html', **results_table)
 
+type_map = {
+    'candidates': lambda x: x,
+    'candidate': map_candidate_page_values,
+    'committees': map_committee_table_values,
+    'committee': map_committee_page_values
+}
+
 def render_page(data_type, c_data):
     # not handling error at api module because sometimes its ok to 
     # not get data back - like with search results
@@ -48,6 +59,6 @@ def render_page(data_type, c_data):
 
     tmpl_vars = c_data['results'][0]
     tmpl_vars.update(type_map[data_type](c_data['results'][0]))
-    tmpl_vars.update(add_committee_data(tmpl_vars, data_type))
+    tmpl_vars.update(add_committee_financial_data(tmpl_vars, data_type))
 
     return render_template(data_type + 's-single.html', **tmpl_vars)
