+++ requirements.txt
@@ -4,6 +4,7 @@ Jinja2==2.7.3
 MarkupSafe==0.23
 Werkzeug==0.9.6
 itsdangerous==0.24
+mock==1.0.1
 nose==1.3.4
 requests==2.5.1
 requests-cache==0.4.9
